backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==23.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv # ==1.0.0
Pillow==9.0.0
webcolors==1.11.1